requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
dateparser==1.1.8
exceptiongroup==1.1.1
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
pyOpenSSL==23.1.1
pytest==7.3.1
python-binance==1.0.17
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.31.0
service-identity==21.1.0
six==1.16.0
tomli==2.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
ujson==5.7.0
urllib3==2.0.2
websocket-client==1.5.2
websockets==11.0.3
yarl==1.9.2
zope.interface==6.0
